Fix: make the -Werror sanitizer/fuzz build clean under gcc 16.
- opt out of -Wmissing-field-initializers: gcc 16 fires it on the
  codebase's designated-init POD idiom for defaulted members.
- rename shadowing geometry-node locals in the mesh reader.
- cast the signed offset index to size_t in the strictness locator.

+++ lib/meios-core/src/meios/urdf/strictness.cpp
@@ -84,7 +84,7 @@ int offset_to_line(std::string_view text, std::ptrdiff_t offset)
     const std::ptrdiff_t stop = std::min<std::ptrdiff_t>(offset, static_cast<std::ptrdiff_t>(text.size()));
     int line = 1;
     for(std::ptrdiff_t i = 0; i < stop; ++i)
-        if(text[i] == '\n')
+        if(text[static_cast<std::size_t>(i)] == '\n')
             ++line;
     return line;
 }
@@ -96,7 +96,7 @@ source_location node_location(pugi::xml_node node, std::string_view text,
     const std::ptrdiff_t stop = std::min<std::ptrdiff_t>(offset, static_cast<std::ptrdiff_t>(text.size()));
     int column = 1;
     for(std::ptrdiff_t i = 0; i < stop; ++i)
-        column = text[i] == '\n' ? 1 : column + 1;
+        column = text[static_cast<std::size_t>(i)] == '\n' ? 1 : column + 1;
     return source_location{ file, offset_to_line(text, offset), column };
 }
